Add permission checks to quote method

+++ src/main/java/me/magandgolden/MainCommand.java
@@ -76,22 +76,54 @@ public class MainCommand extends BaseCommand {
 	@CommandAlias ("quote|quotes")
 	private static void onQuote (CommandSender sender, @Default ("send") String action, @Optional String numberOrText) {
 		Yaml quoteFile = plugin.getQuoteFile();
+		// Create new instnace of dataworks and tells it we are using quoteFile
 		DataWorks dw = new DataWorks(quoteFile);
 		String uuid = (sender instanceof Player) ? ((Player) sender).getUniqueId().toString() : "console";
+
 		switch (action) {
 			case "send":
-				dw.sendRandom(sender);
+				dw.sendRandom(sender); // Does not need to pass list or file bc it was passed to dw above.
 				break;
 			case "add":
-				dw.addFile(sender, uuid, numberOrText);
+				if (! sender.hasPermission("quotes.add")) {  // Check if sender has permission
+					noPermission(sender, "quotes.add");  // if sender does NOT have permission
+					return;// Pass sender and the permission then stop.
+				}
+				if (numberOrText == null) {
+					// If you don't check and it passes null will throw exception
+					sender.sendMessage(ChatColor.RED + "You need to enter something to add");
+					return;
+				}
+				if (! dw.userOverLimit(uuid, "quotes")) // Check how many quotes a user has added
+					dw.addFile(sender, uuid, numberOrText);
+				else {
+					sender.sendMessage(ChatColor.translateAlternateColorCodes('&', // New line is ignored
+							"&fPlease remove a quote before adding more")); // Java continues reading until ;
+					return;
+				}
+				break;
 			case "list":
+				if (! sender.hasPermission("quotes.list")) {
+					noPermission(sender, "quotes.list");
+					return;
+				}
 				dw.sendList(sender);
 				break;
 			case "remove":
+				if (! sender.hasPermission("quotes.remove")) {
+					noPermission(sender, "quotes.remove");
+					return;
+				}
+				// If you don't check and it passes null will throw exception
+				if (numberOrText == null) {
+					sender.sendMessage(ChatColor.RED + "You need to specify what to remove:");
+					dw.sendList(sender);
+					return;
+				}
 				int number = Integer.parseInt(numberOrText);
 				dw.removeFile(sender, number);
-				dw.sendList(sender);
 		}
+
 	}
 
 	private static void noPermission (CommandSender sender, String permission) {
